requirements_decision.txt
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.6.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
asyncio-timeout>=4.0.0